ex18: function pointers + bubble sort
the key idea here is that you can declare a function pointer then use it as an argument to another function, like a callback

when you're finally invoking the function that uses the callback as the argument, you can then pass on the actual name of the defined function.

the key concept here to help you delineate between the two is declare vs define..

this exercise also covers bubble sort which is interesting. mainly
- compare two adjacent values
- if one is bigger, bubble it up to the right (m-1 times))
- you do passes n times, where n is the size of the array